+++ src/main/java/net/geforcemods/securitycraft/inventory/BlockEntityInventoryWrapper.java
@@ -1,12 +1,13 @@
 package net.geforcemods.securitycraft.inventory;
 
+import net.geforcemods.securitycraft.api.IModuleInventoryWithContainer;
 import net.minecraft.world.Container;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.entity.BlockEntity;
 
-public class BlockEntityInventoryWrapper<T extends BlockEntity & Container> implements Container {
+public class BlockEntityInventoryWrapper<T extends BlockEntity & IModuleInventoryWithContainer> implements Container {
 	private final T wrapped;
 	private final AbstractContainerMenu menu;
 
@@ -17,8 +18,11 @@ public class BlockEntityInventoryWrapper<T extends BlockEntity & Container> impl
 
 	@Override
 	public void clearContent() {
-		wrapped.clearContent();
-		menu.slotsChanged(wrapped);
+		for (int i = 0; i < wrapped.getContainerSize(); i++) {
+			wrapped.setContainerItem(i, ItemStack.EMPTY);
+		}
+
+		menu.slotsChanged(this);
 	}
 
 	@Override
@@ -28,44 +32,45 @@ public class BlockEntityInventoryWrapper<T extends BlockEntity & Container> impl
 
 	@Override
 	public boolean isEmpty() {
-		return wrapped.isEmpty();
+		return wrapped.isContainerEmpty();
 	}
 
 	@Override
 	public ItemStack getItem(int index) {
-		return wrapped.getItem(index);
+		return wrapped.getStackInContainer(index);
 	}
 
 	@Override
 	public ItemStack removeItem(int index, int count) {
-		ItemStack toReturn = wrapped.removeItem(index, count);
+		ItemStack toReturn = wrapped.removeContainerItem(index, count, false);
 
-		menu.slotsChanged(wrapped);
+		menu.slotsChanged(this);
 		return toReturn;
 	}
 
 	@Override
 	public ItemStack removeItemNoUpdate(int index) {
-		ItemStack toReturn = wrapped.removeItemNoUpdate(index);
+		ItemStack toReturn = wrapped.getStackInContainer(index);
 
-		menu.slotsChanged(wrapped);
+		wrapped.setStackInSlot(index, ItemStack.EMPTY);
+		menu.slotsChanged(this);
 		return toReturn;
 	}
 
 	@Override
 	public void setItem(int index, ItemStack stack) {
-		wrapped.setItem(index, stack);
-		menu.slotsChanged(wrapped);
+		wrapped.setContainerItem(index, stack);
+		menu.slotsChanged(this);
 	}
 
 	@Override
 	public void setChanged() {
 		wrapped.setChanged();
-		menu.slotsChanged(wrapped);
+		menu.slotsChanged(this);
 	}
 
 	@Override
 	public boolean stillValid(Player player) {
-		return wrapped.stillValid(player);
+		return true;
 	}
 }