+++ src/main/java/net/geforcemods/securitycraft/screen/ProjectorScreen.java
@@ -109,7 +109,7 @@ public class ProjectorScreen extends AbstractContainerScreen<ProjectorMenu> impl
 
 		renderTooltip(guiGraphics, mouseX, mouseY);
 
-		if (slotHoverChecker.checkHover(mouseX, mouseY) && menu.be.isEmpty())
+		if (slotHoverChecker.checkHover(mouseX, mouseY) && menu.be.isContainerEmpty())
 			guiGraphics.renderTooltip(font, slotHoverChecker.getName(), mouseX, mouseY);
 	}